// src/generated/model/content.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Frontend Recruit
 * OpenAPI spec version: 1.0
 */

export interface Content {
  /** コンテンツのID */
  id: number;
  /**
   * コンテンツのタイトル
   * @nullable
   */
  title: string | null;
  /**
   * コンテンツの本文
   * @nullable
   */
  body: string | null;
  /** 作成日時 */
  createdAt: string;
  /** 更新日時 */
  updatedAt: string;
}